# backend/src/main/java/org/roster/backend/adapter/in/web/dto/AuthToken.java
package org.roster.backend.adapter.in.web.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class AuthToken {
    private String token;
}